abstracted the error methods from helpers into their own file
